experiments/demega/requirements.txt
# Core dependencies
torch>=2.0.0
lightning>=2.0.0
torchmetrics>=1.0.0

# Data processing
numpy>=1.24.0
h5py>=3.0.0
pandas>=2.0.0
pyyaml>=6.0

# PNPL library for MEG data loading (September Labs fork)
pnpl @ git+https://github.com/September-Labs/pnpl.git

# HuggingFace datasets for loading preprocessed data
datasets>=2.14.0

# Additional dependencies for PNPL compatibility
mne-bids>=0.14.0
huggingface-hub>=0.20.0
requests>=2.28.0

# Visualization and logging (optional but recommended)
matplotlib>=3.5.0
tensorboard>=2.11.0
wandb>=0.13.0

# Utilities
tqdm>=4.64.0